+++ src/paimon/CMakeLists.txt
@@ -369,6 +369,7 @@ set(PAIMON_CORE_SRCS
     core/realtime/arrow_mem_indexer.cpp
     core/realtime/arrow_mem_indexer_factory.cpp
     core/realtime/realtime_append_only_writer.cpp
+    core/realtime/realtime_context.cpp
     core/realtime/realtime_context_impl.cpp
     core/postpone/postpone_bucket_writer.cpp
     core/schema/arrow_schema_validator.cpp
