+++ src/paimon/core/realtime/realtime_context.cpp
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+#include "paimon/realtime/realtime_context.h"
+
+#include <memory>
+
+#include "paimon/core/realtime/realtime_context_impl.h"
+#include "paimon/realtime/arrow_mem_indexer_factory.h"
+
+namespace paimon {
+
+RealtimeContext::~RealtimeContext() = default;
+
+Result<std::shared_ptr<RealtimeContext>> RealtimeContext::Create() {
+    return Create(std::make_shared<ArrowMemIndexerFactory>());
+}
+
+Result<std::shared_ptr<RealtimeContext>> RealtimeContext::Create(
+    const std::shared_ptr<MemIndexerFactory>& factory) {
+    return RealtimeContextImpl::Create(factory);
+}
+
+}  // namespace paimon
